refactor: validatorV2 제거하고 기존의 validator 사용, 폴더구조 변경
