+++ product-reservation-service/src/main/java/com/monkey/productreservationservice/application/service/ProductReservationServiceApiV2.java
@@ -9,8 +9,8 @@ import com.monkey.productreservationservice.application.dto.response.ResProductR
 import com.monkey.productreservationservice.application.dto.response.ResProductReservationGetDTOApiV1;
 import com.monkey.productreservationservice.application.dto.response.ResProductReservationPostByIdCancelDTOApiV1;
 import com.monkey.productreservationservice.application.dto.response.ResProductReservationPostDTOApiV1;
-import com.monkey.productreservationservice.application.validator.v1.ProductReservationReadValidator;
-import com.monkey.productreservationservice.application.validator.v2.ProductReservationValidatorV2;
+import com.monkey.productreservationservice.application.validator.ProductReservationReadValidator;
+import com.monkey.productreservationservice.application.validator.ProductReservationValidator;
 import com.monkey.productreservationservice.domain.entity.ProductReservationEntity;
 import com.monkey.productreservationservice.domain.repository.ProductReservationRepository;
 import com.monkey.productreservationservice.domain.service.ProductStockServiceV1;
@@ -35,7 +35,7 @@ import java.util.UUID;
 @Slf4j
 public class ProductReservationServiceApiV2 {
     private final ProductReservationRepository productReservationRepository;
-    private final ProductReservationValidatorV2 reservationValidator;
+    private final ProductReservationValidator reservationValidator;
     private final ProductReservationReadValidator readValidator;
     private final ProductFeignClientApiV1 productFeignClientApiV1;
     private final ProductStockServiceV1 productStockService;
